now grabs inner job details from posting url
