--- disney/education_and_qualifications.py
@@ -1,2 +1,0 @@
-#TODO pull job details from links
-
