+++ disney/run.py
@@ -2,8 +2,8 @@
 import pandas as pd
 ####################### My Libraries #####################################
 from scraper import scrape_HTML
-from clean_jobs import separate
-from clean_jobs import split_and_clean
+from filter_job_results import separate
+from clean import split_and_clean
 from open_multiple_links import add_disney_url
 from open_multiple_links import grab_job_data_from_multiple_links
 ############################################################
@@ -19,8 +19,8 @@ if __name__ == "__main__":
     job_details_by_id = separate(job_list)[0]
     paths = separate(job_list)[1]
     links = add_disney_url(paths)
-    # print(links)
-    print(grab_job_data_from_multiple_links(links))
+    job_data = grab_job_data_from_multiple_links(links)
     # create_csv(job_details_by_id)
     # df = create_csv(job_details_by_id)
     # print(df)
+    print(job_data)
