+++ disney/scraper.py
@@ -16,4 +16,5 @@ def launchBrowser():
 def scrape_HTML():
     driver = launchBrowser()
     job_list = driver.find_elements(By.TAG_NAME, value='tbody') #div class that holds job postings
+    # driver.close()
     return job_list
